requirements.txt
streamlit
google-generativeai
deepgram-sdk
python-dotenv
watchdog
pinecone
requests
gTTS